:recycle: refactor: architecture refactoring (#45)
