+++ src/main/java/com/praise/push/adapter/in/web/CommentController.java
@@ -1,0 +1,10 @@
+package com.praise.push.adapter.in.web;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequestMapping("/praise-push/api/v1")
+@RestController
+class CommentController {
+
+}
