+++ src/main/java/com/praise/push/application/port/in/KeywordUseCase.java
@@ -1,0 +1,12 @@
+package com.praise.push.application.port.in;
+
+import com.praise.push.domain.Keyword;
+
+import java.util.List;
+
+public interface KeywordUseCase {
+    /**
+     * returns random keywords as many as size
+     */
+    List<Keyword> getRandomRecommendationKeywords(Integer size);
+}
