+++ src/main/java/com/praise/push/domain/Keyword.java
@@ -1,0 +1,20 @@
+package com.praise.push.domain;
+
+import lombok.*;
+
+@Builder
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class Keyword {
+
+    /**
+     * keyword's id
+     */
+    private Long id;
+
+    /**
+     * keyword's value
+     */
+    private String keyword;
+}
