--- src/main/java/com/praise/push/post/application/port/in/CreatePostCommand.java
@@ -1,15 +1,0 @@
-package com.praise.push.post.application.port.in;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class CreatePostCommand {
-    private String title;
-    private String content;
-    private String imageUrl;
-    private String keyword;
-}
