--- src/main/java/com/praise/push/post/application/port/in/KeywordUseCase.java
@@ -1,12 +1,0 @@
-package com.praise.push.post.application.port.in;
-
-import com.praise.push.post.domain.Keyword;
-
-import java.util.List;
-
-public interface KeywordUseCase {
-    /**
-     * returns random keywords as many as size
-     */
-    List<Keyword> getRandomRecommendationKeywords(Integer size);
-}
