--- src/main/java/com/praise/push/post/domain/Keyword.java
@@ -1,20 +1,0 @@
-package com.praise.push.post.domain;
-
-import lombok.*;
-
-@Builder
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class Keyword {
-
-    /**
-     * keyword's id
-     */
-    private Long id;
-
-    /**
-     * keyword's value
-     */
-    private String keyword;
-}
